[test] 테스트 코드 내 config type 통일화

+++ src/test/java/com/github/cokelee777/kafka/connect/smt/claimcheck/source/S3ClaimCheckSourceTransformTest.java
@@ -234,7 +234,7 @@ public class S3ClaimCheckSourceTransformTest {
       }
 
       private void assertClaimChecked(
-          SourceRecord originalRecord, SourceRecord transformedRecord, int originalSizeBytes) {
+          SourceRecord originalRecord, SourceRecord transformedRecord, long originalSizeBytes) {
         assertNotSame(originalRecord, transformedRecord);
         assertEquals(ClaimCheckSchema.SCHEMA, transformedRecord.valueSchema());
